requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
requests==2.31.0

chromadb==0.4.5

sentence-transformers==2.2.2
numpy==1.26.4

openai==1.31.0
pydantic==2.6.1